feat: set ID and description for VIPs 434 and 436
