+++ simulations/vip-436/bscmainnet.ts
@@ -1,11 +1,11 @@
 import { expectEvents } from "src/utils";
 import { forking, testVip } from "src/vip-framework";
 
-import vip433 from "../../vips/vip-433/bscmainnet";
+import vip436 from "../../vips/vip-436/bscmainnet";
 import OMNICHAIN_PROPOSAL_SENDER_ABI from "./abi/OmnichainProposalSender.json";
 
 forking(46110084, async () => {
-  testVip("VIP-433", await vip433({}), {
+  testVip("VIP-436", await vip436({}), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(
         txResponse,
