+++ simulations/vip-436/bsctestnet.ts
@@ -1,11 +1,11 @@
 import { expectEvents } from "src/utils";
 import { forking, testVip } from "src/vip-framework";
 
-import vip433 from "../../vips/vip-433/bsctestnet";
+import vip436 from "../../vips/vip-436/bsctestnet";
 import OMNICHAIN_PROPOSAL_SENDER_ABI from "./abi/OmnichainProposalSender.json";
 
 forking(47754551, async () => {
-  testVip("VIP-433", await vip433(), {
+  testVip("VIP-436", await vip436(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(
         txResponse,
