+++ vips/vip-434/bscmainnet.ts
@@ -10,7 +10,7 @@ const { POOL_REGISTRY, VTREASURY, NORMAL_TIMELOCK, CHAINLINK_ORACLE, RESILIENT_O
 
 export const COMPTROLLER_CORE = "0x317c1A5739F39046E20b08ac9BeEa3f10fD43326";
 export const CHAINLINK_STALE_PERIOD = 26 * 60 * 60; // 26 hours
-export const gmBTC_CHAINLINK_FEED = "0x395D5c5D552Df670dc4B2B1cef0c4EABfFba492f";
+export const gmWETH_CHAINLINK_FEED = "0xfB3264D1129824933a52374c2C1696F4470D041e";
 
 export const USDT_PRIME_CONVERTER = "0x435Fac1B002d5D31f374E07c0177A1D709d5DC2D";
 export const USDC_PRIME_CONVERTER = "0x6553C9f9E131191d4fECb6F0E73bE13E229065C6";
@@ -27,8 +27,8 @@ export const BaseAssets = [
 const CONVERSION_INCENTIVE = parseUnits("0.0001", 18);
 
 export const REFUND_ADDRESS = "0xe1f7c5209938780625E354dc546E28397F6Ce174";
-export const REFUND_AMOUNT = parseUnits("10000", 18);
-export const REFUND_TOKEN = "0x7C11F78Ce78768518D743E81Fdfa2F860C6b9A77";
+export const REFUND_AMOUNT = parseUnits("11000", 18);
+export const REFUND_TOKEN = "0x450bb6774Dd8a756274E0ab4107953259d2ac541";
 
 type Token = {
   address: string;
@@ -37,7 +37,7 @@ type Token = {
 };
 
 export const token = {
-  address: "0x47c031236e19d024b42f8AE6780E44A573170703",
+  address: "0x70d95587d40A2caf56bd97485aB3Eec10Bee6336",
   decimals: 18,
   symbol: "GM",
 };
@@ -75,9 +75,9 @@ type Market = {
 
 export const market: Market = {
   vToken: {
-    address: "0x4f3a73f318C5EA67A86eaaCE24309F29f89900dF",
-    name: "Venus gmBTC-USDC (Core)",
-    symbol: "vgmBTC-USDC_Core",
+    address: "0x9bb8cEc9C0d46F53b4f2173BB2A0221F66c353cC",
+    name: "Venus gmWETH-USDC (Core)",
+    symbol: "vgmWETH-USDC_Core",
     underlying: token,
     decimals: 8,
     exchangeRate: parseUnits("1", 28),
@@ -86,29 +86,83 @@ export const market: Market = {
   riskParameters: {
     collateralFactor: parseUnits("0.55", 18),
     liquidationThreshold: parseUnits("0.6", 18),
-    supplyCap: parseUnits("2650000", 18),
+    supplyCap: parseUnits("2000000", 18),
     borrowCap: parseUnits("0", 18),
     reserveFactor: parseUnits("0.25", 18),
     protocolSeizeShare: parseUnits("0.05", 18),
   },
   initialSupply: {
-    amount: parseUnits("4800", 18),
+    amount: parseUnits("6000", 18),
     vTokenReceiver: "0xe1f7c5209938780625E354dc546E28397F6Ce174",
   },
   interestRateModel: {
-    address: "0x390D1C248217615D79f74f2453D682906Bd2dD20",
+    address: "0x425dde630be832195619a06175ba45C827Dd3DCa",
     base: "0",
-    multiplier: "0.15",
-    jump: "2.5",
-    kink: "0.45",
+    multiplier: "0.03",
+    jump: "4.5",
+    kink: "0.9",
   },
 };
 
 const vip434 = (overrides: { chainlinkStalePeriod?: number }) => {
   const meta = {
     version: "v2",
-    title: "VIP-433",
-    description: ``,
+    title: "VIP-434 [Arbitrum] New gmETH market in the Core pool",
+    description: `#### Summary
+
+If passed, following the Community proposal “[Add gmETH on Venus Protocol](https://community.venus.io/t/add-gmeth-on-venus-protocol/4767)” and [the associated snapshot](https://snapshot.org/#/s:venus-xvs.eth/proposal/0x84e11121a9bf417ae20261798c0f30f8a2da050002db20a4e46a62ffa3dd8346), this VIP adds a market for [gmETH (WETH-USDC)](https://arbiscan.io/address/0x70d95587d40A2caf56bd97485aB3Eec10Bee6336) into the Core pool on Arbitrum one.
+
+Moreover, this VIP would transfer 11,000 [gmETH (WETH-WETH)](https://arbiscan.io/address/0x450bb6774Dd8a756274E0ab4107953259d2ac541) tokens from the [Venus Treasury](https://arbiscan.io/address/0x8a662ceac418daef956bc0e6b2dd417c80cda631) to the [GMX.io](https://arbiscan.io/address/0xe1f7c5209938780625E354dc546E28397F6Ce174) project. They [provided these tokens](https://arbiscan.io/tx/0x881d12517bee1b9e86553977985036ac87d3875b341d0ef2ca11952c10927704), but they are now unnecessary to launch the new market.
+
+#### Description
+
+**Risk parameters**
+
+Following [Chaos Labs recommendations](https://community.venus.io/t/add-gmeth-on-venus-protocol/4767/7), the risk parameters for the new market are:
+
+Underlying token: [gmETH (WETH-USDC)](https://arbiscan.io/address/0x70d95587d40A2caf56bd97485aB3Eec10Bee6336)
+
+- Borrow cap: 0 gmETH (WETH-USDC)
+- Supply cap: 2,000,000 gmETH (WETH-USDC)
+- Collateral factor: 55%
+- Liquidation threshold: 60%
+- Reserve factor: - (not relevant because the asset won’t be borrowable)
+
+Bootstrap liquidity: 6,000 gmETH (WETH-USDC) provided by [GMX.io](https://arbiscan.io/address/0xe1f7c5209938780625E354dc546E28397F6Ce174).
+
+The interest rate curve for the new market is not relevant because the asset is not borrowable, but these parameters will be set anyway:
+
+- kink: 90%
+- base (yearly): 0%
+- multiplier (yearly): 3%
+- jump multiplier (yearly): 450%
+
+#### Oracles configuration
+
+The [ResilientOracle](https://docs-v4.venus.io/risk/resilient-price-oracle) deployed to [Arbitrum one](https://arbiscan.io/address/0xd55A98150e0F9f5e3F6280FC25617A5C93d96007) is used for gmETH (WETH-USDC), using under the hood the Chainlink price ([feed](https://arbiscan.io/address/0xfB3264D1129824933a52374c2C1696F4470D041e)).
+
+#### Security and additional considerations
+
+We applied the following security procedures for this upgrade:
+
+- **VIP execution simulation**: in a simulation environment, validating the new market is properly added to the Core pool on Arbitrum one, with the right parameters and the expected bootstrap liquidity
+- **Deployment on testnet**: the same market has been deployed to Arbitrum sepolia, and used in the Venus Protocol testnet deployment
+
+#### Deployed contracts
+
+Mainnet
+
+- vgmWETH-USDC_Core: [0x9bb8cEc9C0d46F53b4f2173BB2A0221F66c353cC](https://arbiscan.io/address/0x9bb8cEc9C0d46F53b4f2173BB2A0221F66c353cC)
+
+Sepolia
+
+- vgmWETH-USDC_Core: [0x4A80b19Cd8BbBd14c425fB17F8E06c6B60801d63](https://arbiscan.io/address/0x4A80b19Cd8BbBd14c425fB17F8E06c6B60801d63)
+
+#### References
+
+- [VIP simulation](https://github.com/VenusProtocol/vips/pull/468)
+- [Deployment to Arbitrum Sepolia](https://sepolia.arbiscan.io/tx/0x155de4290f1cbe187680d81440469b51563dbf2087331390eac7e30b75dd4d21)
+- [Documentation](https://docs-v4.venus.io/)`,
     forDescription: "I agree that Venus Protocol should proceed with this proposal",
     againstDescription: "I do not think that Venus Protocol should proceed with this proposal",
     abstainDescription: "I am indifferent to whether Venus Protocol proceeds or not",
@@ -121,7 +175,7 @@ const vip434 = (overrides: { chainlinkStalePeriod?: number }) => {
       {
         target: CHAINLINK_ORACLE,
         signature: "setTokenConfig((address,address,uint256))",
-        params: [[token.address, gmBTC_CHAINLINK_FEED, chainlinkStalePeriod]],
+        params: [[token.address, gmWETH_CHAINLINK_FEED, chainlinkStalePeriod]],
         dstChainId: LzChainId.arbitrumone,
       },
       {
